+++ src/apps/sequencer/engine/CurveTrackEngine.cpp
@@ -240,11 +240,14 @@ void CurveTrackEngine::updateOutput(uint32_t relativeTick, uint32_t divisor) {
 }
 
 bool CurveTrackEngine::isRecording() const {
-    return
+    bool val =
         _engine.state().recording() &&
         _curveTrack.curveCvInput() != Types::CurveCvInput::Off;
-        //&&
-        //_model.project().selectedTrackIndex() == _track.trackIndex();
+
+    if (!_model.project().useMultiCvRec()) {
+        return val && _model.project().selectedTrackIndex() == _track.trackIndex();
+    }
+    return val;
 }
 
 void CurveTrackEngine::updateRecordValue() {