+++ src/apps/sequencer/model/Project.cpp
@@ -43,6 +43,7 @@ void Project::clear() {
     setCvGateInput(Types::CvGateInput::Off);
     setCurveCvInput(Types::CurveCvInput::Off);
     setResetCvOnStop(true);
+    setUseMultiCvRec(true);
 
     _clockSetup.clear();
 
@@ -113,9 +114,9 @@ void Project::write(VersionedSerializedWriter &writer) const {
     writer.write(_monitorMode);
     writer.write(_recordMode);
     writer.write(_midiInputMode);
+    _midiInputSource.write(writer);
     writer.write(_midiIntegrationMode);
     writer.write(_midiProgramOffset);
-    _midiInputSource.write(writer);
     writer.write(_cvGateInput);
     writer.write(_curveCvInput);
 
@@ -135,6 +136,7 @@ void Project::write(VersionedSerializedWriter &writer) const {
     writer.write(_selectedTrackIndex);
     writer.write(_selectedPatternIndex);
     writer.write(_resetCvOnStop);
+    writer.write(_useMultiCv);
 
     writer.writeHash();
 
@@ -162,13 +164,10 @@ bool Project::read(VersionedSerializedReader &reader) {
     }
     if (reader.dataVersion() >= ProjectVersion::Version32) {
         reader.skip<bool>(ProjectVersion::Version32, ProjectVersion::Version38);
-        setMidiIntegrationMode(Types::MidiIntegrationMode::None);
-        setMidiProgramOffset(0);
-    }
-    if (reader.dataVersion() >= ProjectVersion::Version38) {
         reader.read(_midiIntegrationMode);
         reader.read(_midiProgramOffset);
     }
+
     reader.read(_cvGateInput, ProjectVersion::Version6);
     reader.read(_curveCvInput, ProjectVersion::Version11);
 
@@ -190,6 +189,7 @@ bool Project::read(VersionedSerializedReader &reader) {
     reader.read(_selectedTrackIndex);
     reader.read(_selectedPatternIndex);
     reader.read(_resetCvOnStop, ProjectVersion::Version38);
+    reader.read(_useMultiCv, ProjectVersion::Version39);
 
     bool success = reader.checkHash();
     if (success) {