+++ src/apps/sequencer/ui/controllers/launchpad/LaunchpadController.cpp
@@ -596,7 +596,8 @@ void LaunchpadController::sequenceButton(const Button &button, ButtonAction acti
                     }
                     auto &sequence = _project.selectedStochasticSequence();
                     sequence.step(selectedNote).toggleGate();
-                    sequence.step(selectedNote).setNoteOctave(selectedOctave);                }
+                    sequence.step(selectedNote).setNoteOctave(selectedOctave);                
+                    }
                 break;
             }
             case Track::TrackMode::Arp: {
@@ -630,7 +631,7 @@ void LaunchpadController::sequenceButton(const Button &button, ButtonAction acti
 }
 
 void LaunchpadController::manageCircuitKeyboard(const Button &button) {
-    const auto &sequence = _project.selectedNoteSequence();
+    auto &sequence = _project.selectedNoteSequence();
     const auto &scale = sequence.selectedScale(_project.scale());
     const Scale &bypasssScale = Scale::get(0);
     switch ( _project.selectedNoteSequenceLayer()) {
@@ -717,6 +718,7 @@ void LaunchpadController::manageCircuitKeyboard(const Button &button) {
             } else if (button.row == 7) {
                 if (button.col <=3) {
                     Button btn = Button(3,button.col);
+                    sequence.setSecion(button.col);
                     navigationButtonDown(_sequence.navigation, btn);  
                 }
             }
@@ -829,7 +831,7 @@ void LaunchpadController::manageStochasticCircuitKeyboard(const Button &button) 
                 break;
             }
         default:
-            sequenceEditStep(button.row, button.col);
+            //sequenceEditStep(button.row, button.col);
             break;
         break;
     }   
@@ -929,7 +931,7 @@ void LaunchpadController::manageArpCircuitKeyboard(const Button &button) {
                 break;
             }
         default:
-            sequenceEditStep(button.row, button.col);
+            //sequenceEditStep(button.row, button.col);
             break;
         break;
     }