+++ src/apps/sequencer/ui/pages/ArpSequencePage.cpp
@@ -67,7 +67,6 @@ void ArpSequencePage::updateLeds(Leds &leds) {
 void ArpSequencePage::keyPress(KeyPressEvent &event) {
     const auto &key = event.key();
 
-      functionShortcuts(event);
 
     if (key.shiftModifier() && event.count() == 2) {
         saveContextShow();