+++ src/apps/sequencer/ui/pages/NoteSequenceEditPage.cpp
@@ -364,8 +364,6 @@ void NoteSequenceEditPage::keyPress(KeyPressEvent &event) {
     auto &sequence = _project.selectedNoteSequence();
     auto &track = _project.selectedTrack().noteTrack();
 
-    functionShortcuts(event);
-
     auto &trackEngine = _engine.selectedTrackEngine().as<NoteTrackEngine>();
 
     if (key.isContextMenu()) {