+++ src/apps/sequencer/ui/pages/StochasticSequenceEditPage.cpp
@@ -386,8 +386,6 @@ void StochasticSequenceEditPage::keyPress(KeyPressEvent &event) {
     const auto &key = event.key();
     auto &sequence = _project.selectedStochasticSequence();
 
-    functionShortcuts(event);
-
     if (key.isContextMenu()) {
         contextShow();
         event.consume();